acp_decoders/requirements.txt
wheel
paho-mqtt
simplejson
gmqtt
uvloop
